requirements.txt
Flask==2.2.2
gunicorn==20.1.0
dataclasses==0.6
dataclasses-json==0.5.7
google-cloud==0.34.0
google-cloud-language==2.6.0
protobuf==4.21.7
datasets==2.5.1
scikit-learn==1.1.2
evaluate==0.2.2
